pyproject.toml
[tool.poetry]
name = "docper"
version = "0.1.0"
description = ""
authors = ["Zohre <tabatabaee.zohre@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.20.0"
fastapi = "^0.88.0"
tortoise-orm = "^0.19.2"
asyncpg = "^0.27.0"
gunicorn = "^20.1.0"
pytest = "^7.2.0"
requests = "^2.28.2"
httpx = "^0.23.3"
pytest-asyncio = "^0.20.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
